Conflicts:
	conf/susi/en_0001_foundation.txt

+++ src/ai/susi/mind/SusiMemory.java
@@ -78,7 +78,7 @@ public class SusiMemory {
     public SusiMemory(File chatlog, File skilllog, int attention) {
         this.chatlog = chatlog;
         this.skilllog = skilllog;
-        try {FileUtils.cleanDirectory(this.skilllog);} catch (IOException e) {} // do this only as long as we are in a migration phase
+        if (this.skilllog != null) try {FileUtils.cleanDirectory(this.skilllog);} catch (IOException e) {} // do this only as long as we are in a migration phase
         this.attention = attention;
         this.memories = new ConcurrentHashMap<>();
         this.intentsets = new ConcurrentHashMap<>();
